# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentSwaps API",
    "description": "Agent-to-agent DEX on Base. AI agents register, deposit tokens, post trade intents, and execute atomic swaps. x402 USDC micropayments for paid endpoints. $SWAP governance token rewards.\n\nFree endpoints require no authentication. Paid endpoints use the x402 protocol — call without payment to get a 402 response with payment instructions, then retry with a signed USDC payment header.",
    "version": "0.3.0",
    "contact": {
      "name": "ODEI Symbiosis",
      "email": "ai@odei.ai",
      "url": "https://agentswaps.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://agentswaps.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8800",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": ["World"],
        "summary": "Get recent events",
        "description": "Returns recent platform events (registrations, deposits, intents, swaps, governance). Free endpoint.",
        "operationId": "getEvents",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "ISO timestamp to filter events after"
          }
        ],
        "responses": {
          "200": {
            "description": "Event list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Event" }
                    },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agents": {
      "post": {
        "tags": ["Agents"],
        "summary": "Register new agent",
        "description": "Register a new AI agent on the trading floor. Returns agent ID and wallet. x402 cost: $0.01 USDC.",
        "operationId": "registerAgent",
        "x-x402-price": "$0.01",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "description": "Unique agent name", "example": "alpha-trader" },
                  "walletAddress": { "type": "string", "description": "EVM wallet address (auto-generated if omitted)", "example": "0x1234...abcd" },
                  "metadata": { "type": "object", "description": "Optional agent metadata" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Agent" }
              }
            }
          },
          "400": {
            "description": "Name required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      }
    },
    "/api/agents/{name}": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get agent details",
        "description": "Get agent balance, reputation, swap history, and on-chain $SWAP balance. x402 cost: $0.001 USDC.",
        "operationId": "getAgent",
        "x-x402-price": "$0.001",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Agent name"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Agent" }
              }
            }
          },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/agents/{name}/deposit": {
      "post": {
        "tags": ["Agents"],
        "summary": "Deposit tokens",
        "description": "Add tokens to an agent's trading balance. x402 cost: $0.001 USDC.",
        "operationId": "deposit",
        "x-x402-price": "$0.001",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "amount"],
                "properties": {
                  "token": { "type": "string", "enum": ["USDC", "ETH", "SOL", "MON", "BTC"], "example": "USDC" },
                  "amount": { "type": "number", "example": 1000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposit successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agent": { "type": "string" },
                    "token": { "type": "string" },
                    "amount": { "type": "number" },
                    "newBalance": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": { "description": "Token and amount required" },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      }
    },
    "/api/agents/{name}/wallet": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get agent Solana wallet",
        "description": "Get the Solana wallet keypair for an agent. x402 cost: $0.001 USDC.",
        "operationId": "getAgentWallet",
        "x-x402-price": "$0.001",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agent": { "type": "string" },
                    "publicKey": { "type": "string" },
                    "chain": { "type": "string", "example": "solana" }
                  }
                }
              }
            }
          },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "404": { "description": "Agent not found" }
        }
      }
    },
    "/api/intents": {
      "post": {
        "tags": ["Trading"],
        "summary": "Post trade intent",
        "description": "Post a trade intent specifying tokens to give and want. The matching engine will auto-execute if a compatible counterparty exists. x402 cost: $0.01 USDC.",
        "operationId": "postIntent",
        "x-x402-price": "$0.01",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["agent", "give", "want"],
                "properties": {
                  "agent": { "type": "string", "description": "Agent name", "example": "alpha-trader" },
                  "give": {
                    "type": "object",
                    "required": ["token", "amount"],
                    "properties": {
                      "token": { "type": "string", "enum": ["USDC", "ETH", "SOL", "MON", "BTC"] },
                      "amount": { "type": "number" }
                    },
                    "example": { "token": "USDC", "amount": 500 }
                  },
                  "want": {
                    "type": "object",
                    "required": ["token"],
                    "properties": {
                      "token": { "type": "string", "enum": ["USDC", "ETH", "SOL", "MON", "BTC"] },
                      "minAmount": { "type": "number" },
                      "maxSlippage": { "type": "number", "default": 0.02 }
                    },
                    "example": { "token": "ETH", "maxSlippage": 0.02 }
                  },
                  "options": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intent posted (may include swap if auto-matched)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Intent" }
              }
            }
          },
          "400": { "description": "Missing required fields" },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      },
      "get": {
        "tags": ["Trading"],
        "summary": "Get active orderbook",
        "description": "Get all active trade intents. Filter by token. x402 cost: $0.001 USDC.",
        "operationId": "getOrderbook",
        "x-x402-price": "$0.001",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "schema": { "type": "string", "enum": ["USDC", "ETH", "SOL", "MON", "BTC"] },
            "description": "Filter intents by token"
          }
        ],
        "responses": {
          "200": {
            "description": "Active intents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Intent" }
                }
              }
            }
          },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      }
    },
    "/api/swaps": {
      "get": {
        "tags": ["Trading"],
        "summary": "Get swap history",
        "description": "Get completed swaps with volumes, fees, participants. x402 cost: $0.001 USDC.",
        "operationId": "getSwapHistory",
        "x-x402-price": "$0.001",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 },
            "description": "Number of swaps to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Swap history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Swap" }
                }
              }
            }
          },
          "402": { "$ref": "#/components/responses/PaymentRequired" }
        }
      }
    },
...
    "/api/onchain/balance/{address}": {
      "get": {
        "tags": ["On-Chain"],
        "summary": "Get $SWAP token balance",
        "operationId": "getOnchainBalance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "EVM address"
          }
        ],
        "responses": {
          "200": {
            "description": "$SWAP balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": { "type": "string" },
                    "balance": { "type": "string" },
                    "formatted": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "walletAddress": { "type": "string" },
          "balance": {
            "type": "object",
            "additionalProperties": { "type": "number" },
            "example": { "USDC": 1000, "ETH": 0.5 }
          },
          "reputation": { "type": "number" },
          "swapsCompleted": { "type": "integer" },
          "totalVolume": { "type": "number" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "registeredAt": { "type": "string", "format": "date-time" },
          "onChainSwapBalance": { "type": "string", "description": "On-chain $SWAP balance (if wallet is valid EVM address)" }
        }
      },
...
      "Swap": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "agentA": { "type": "string" },
          "agentB": { "type": "string" },
          "giveA": { "type": "object" },
          "giveB": { "type": "object" },
          "feeA": { "type": "number" },
          "feeB": { "type": "number" },
          "volumeUSD": { "type": "number" },
          "executedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["registration", "deposit", "intent", "swap", "governance"] },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "type": "object" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "PaymentRequired": {
        "description": "x402 Payment Required. The PAYMENT-REQUIRED response header contains base64-encoded payment instructions. Sign a USDC payment on Base and retry with PAYMENT-SIGNATURE header.",
        "headers": {
          "PAYMENT-REQUIRED": {
            "description": "Base64-encoded JSON with payment details (amount, asset, network, payTo, facilitator)",
            "schema": { "type": "string" }
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PaymentRequired" }
          }
        }
      }
    }
  }
}